LibSoftGPU: Move enums into separate file
